mindguard-ai/backend/requirements.txt
fastapi
uvicorn[standard]
google-generativeai
pydantic
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
python-multipart
motor
pymongo
python-dotenv
